requirements.txt
anyio==4.3.0
asgiref==3.8.1
Babel==2.14.0
black==24.3.0
boto3==1.34.79
botocore==1.34.79
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-soft-delete==1.0.12
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
mock==5.1.0
model-bakery==1.17.0
mypy-extensions==1.0.0
oauth2client==4.1.3
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.34
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
py-moneyed==3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
structlog==24.1.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.43.0
whitenoise==6.6.0
